feat: Collect and use zigfigs to make Zigs
Instead of relying only on somewhat buggy plans to enter the Depths Zig,
collect ziggurat figurines and use them when they're available. This
includes a simple plan to move to an adjacent square if we can't use a
figurine on our current square. This would have to be expanded to
properly support moving to anywhere in LOS to support using multiple
figurines for megazigging, but qw is a ways from being able to clear
megazigs, so this will do for now.