feat: Split lua into multiple files
Working with a single 10k line lua file is bad for a number of reasons.
This commit replaces qw.lua with a source directory containing
individual lua files. I've split up the functions into respective files
as best I can, with plans going into a set of files having the prefix
`plans-`. Future commits will likely change files and data around based
on which makes the most sense for qw's development.

In order for a multiple file setup to work without having to define a
specific file order so that all the plans in order of reference by
parent cascade plan, all cascade plans are defined as local variables
that are only initialized at the initialization stage, which is after
all functions have been defined.

The make-qw-rc.sh script is renamed to make-qw.sh and assembles a single
lua file only by default, but can also still add the contents of this
lua file to an rc with its `-r` argument. The `variables.lua` file must
be the first file included so that all local variables are available as
upvalues for functions using them. In the future it'd be good to make to
a proper makefile instead of using this custom make script.