feat: Some iterators for LOS
Instead of manually setting up double for loops whenever we want to
iterate over LOS, make a proper generator-style interator for
rectangular and radial LOS iteration. Also add an adjacent squares
iterator based on the rectangular iterator with radius 1. The radius
iterator starts from the center and iterators outwards by radius. This
is a more complicated iterator and doesn't yet have a use case. It may
be helpful for optimizing iterations where we'll always want to stop at
the closest set of mosnters, so I'm adding it for now.