fix: Use better lua inlining
Currently our lua must add some kind of text on lines containing a sole
`}`, otherwise this terminates our rcfile lua block. This commit
switches `make-qw.sh` to using `<` and `>` to inline our lua, which
removes the need for these `--hack` comments. It's easy to trigger
errors that are either confusing or completely silent by forgetting to
consider lines with only `{`, but with this commit, that's no longer a
consideration

This change is possible since our rcfile variables don't depend on any
functions we've defined in lua. We should be able to continue to avoid
the need for a dependency like that, but hopefullly even if one arises
we can use a decidated lua file that does use `{` and `}` to inline.