feat: Tweaks to default gameplan lists
Don't explore all of Slime in the 15 rune and All plans, just dive to
Slime:5 and fully explore that level. For all plans except 3 runes, do
only Zot:1-4 before attempting god conversion and the rest of extended.
This makes us a bit stronger in terms of XP without subjecting us to the
heavy mutations and general resource consumption of Zot:5.