fix: Travel to and from Hell branches
Recent improvements to the travel system have broken travel in hell
branches under certain conditions. One aspect of this was that travel
always walked a branch's parent chain even when the destination branch
was the same as our current branch. For Hell branches, we only need to
do this if we're considering an in-branch destination above us, since
we'd have to take the exit to Hell. Otherwise there's no need to
consider parent branches for in-branch travel. Travel was also not aware
that it could always reach Hell from a Hell branch by taking the
guaranteed exit on every floor. This commit fixes travel by making the
travel search aware of these aspects.

One thing we should do that's not currently done is track branch exits
like we do branch entries. This would mostly be relevant for accurate
travel in Hell branches and wizmode. Currently travel assumes the exit
is always reachable from a hell branch level, but that may not be the
case when a level isn't fully explored.