feat: Better gameplan handling of god worship
Previously we just relied on the plan cascade to get us to an altar when
we wanted to join a god. But certain conditions like the precense of
runed doors on levels with viable altars can mess with these plans by
repeatedly allowing autoexplore to trigger. This commit has the gameplan
system recognize when we're ready to join a religion, updating the
gameplan status accordingly. This way autoexplore will probably get
disabled, since the gameplan system will see that we want to do a stash
search travel to the altar.