feat: Quit if we get to 1 max hp
Now that the Zot clock doesn't outright kill you, it's possible for qw
to go on forever stuck in a loop in various cases that don't involve it
actually determining that it's stuck. For example, it may explore or
travel towards a area with monsters behind deep water and retreat once
taking damage, only to rest to full and make the attempt again. Future
functionality will make qw better able to deal with all such scenarious,
but intil then we need qw to eventually quit when runs into the Zot
clock in these situations. This commit has qw always quit if it reaches
a max hp of 1.