feat: Tweak teleporting when stuck
Currently qw tries to teleport at the end of its movement plan, when
it's stuck, before trying random movement. This can waste valuable
teleport scrolls in situations where qw is very temporarily stuck. This
commit puts a random movement plan before the teleport plan that's only
active if we've been stuck less than 50 turns. After 50 turns of still
being stuck, qw will attempt it's plan to teleport. The final plan of
the movement cascade is still the main random movement plan, which lasts
for 1000 turns total before qw quits the game.

We reset the stuck counter whenever a level transition is made so that
this approach can work. There was no reason to make it cumulative across
a session in the first place, since when qw is stuck, it's never moving
between levels.