fix: Allow following specialized rune level plans
With recent changes to stair exploration in a18a5b11, branch-specific
plans for navigating to the rune for Swamp were no longer properly
triggered. Instead the level exploration code was resetting stair
exploration, which made it hard for qw to make much progress, since it
had to travel up and try stairs before proceeding further into the cloud
generators. This commit has the exploration code not perform a reset if
we've autoexplored the rune level and found all upstairs, which will
allow the Swamp plans to take over until (hopefully) the rune is found.

Also fix the stash search travel to the rune that we broke some time ago
by omitting a space in our search.