fix: Issues with stair exploration
This commit is a significant rewrite of stair-related aspects of travel.
Although qw doesn't get stuck often these days with the recent
improvements to exploration and travel, this case in Orc revealed that
it still has some blind spots. Specific qw got shafted from Orc:1,
partially explored Orc:2 to reveal two upstairs, and got stuck in a loop
between Orc:1 and Orc:2 when the upstairs it took lead to a disconnected
cell with no further stairs. In this case the necessary backtracking was
broken as a result of tight control travel has now over unexplored stair
searches.

The code to determine the next travel depth and associated direction was
relatively difficult to understand in terms of it repeating various bits
of logic several times in the relevant function. This commit refactors
and simplifies the logic into fewer functions with more general code. It
removes special state tracking for stairs searches that's made redundant
by aforementioned recent changes to stairs direction searches. We also
no longer use an explicit backtracking state variable, instead relying
on stair state resets and the backtracking plan to put us in a state
that has us search new stairs on the level to which we backtrack.