refactor: Use iterators for LOS
Use the previously added iterators instead of constructing indent-heavy
double loops for iterating over some radius. This removes two levels of
indent in some cases where we previously had to test for not being at
the center, since the iterators do this by default.

We also now set a default radius in these iterators to the current LOS
radius.