fix: Issues related to god and branch planning
Correctly set the planning variables for TSO and add a new one for the
good gods in general. The latter is used for equipment checks for items
hated by all good gods.

For god planning variables, set these planning variables to true if our
current god does or any future god will meet the relevant criteria. For
branch planning variables, we still only set them to true if our current
or future plans include the relevant branch. This way we can do
equipment swaps while in the branch of the previous (now complete) plan
and before we enter our new branch. The exception to this is Pan, since
it can take a while to exit that branch after we find all the runes, and
we'll always have an opportunity to swap equipment based on new plans
after we exit Pan.

For Zigs, remove a secondary planning variable that we no longer need.