fix: Update LOS radius values and refactor
Move the LOS radius definition to player.lua and create a function to
initialize it. Currently this has per-species value, which now includes
the correct value for Kobolds, so it's called only during
initialization. In the future we'd like to account for scarf of shadows,
Ring of Shadows, and Robe of Night and have a scheme that updates it
when these items are equipped. Either that or expose the los radius via
the clua API. For now, this will do.