feat: Expand distance mapping functionality
As of crawl/crawl@d42a0f4d, we can use waypoints on all levels. This
means we can create a traversability map for the entire level on every
level unconditionally, even in Abyss. By generalizing code previously
only used for distance mapping to upstairs, this commit takes advantage
of the waypoint changes and adds a general framework for calculating
distance maps to any type of feature or any specific location.

The travel gets some changes to allow caching the first direction we'd
like to take for our next travel destination, which lets us use the new
distance mapping functionality whenever we're unable to autotravel.
Currently this functionality only resolves some situations involving
monsters that are unreachable according to LOS_NO_TRANS but are
reachable via areas connected to our current position. In these cases,
we'll be able to reach a desired staircase or the monster itself.

The distance mapping functionality needs expansion to handle various
situations where we need to attack or exclude hostile monsters in
unreachable areas. This will need to involve things like monster
assessment, the availability of ranged options, and the management of
exclusions.

We also add a debug variable `USE_COROUTINE` that can be toggled with
the `toggle_coroutine()` function. Setting this to false avoids use of
the coroutine, which can be helpful in testing.

A lot of variables have been renamed for consistency with respect to
their purpose, including some not directly related to the traversability
map. I've moved map-centric code to a new `source/map.lua` file from its
previous home in `source/stairs.lua`.