feat: Move qw fully into a coroutine
We're going to be performing expensive floodfill calculations
potentially during the plan cascade and need to be able to break them up
across `ready()` calls to help avoid throttling. This commit puts what
was previously the ready function into a coroutine.

As qw becomes more computationally intensive, we may need both
bot-specific builds that remove the lua throttle as well as more yield
points. The latter will allow the coroutine approach to be more
effective, however it may take some profiling to determine where yield
points need to be placed.