feat: The 0.2 feature branches merge
This commit merges the contents of the floodfilll and ranged-combat
branches in one commit. Ideally, I'd have split this into a nice set of
logical commits. However, with almost 100 "work in progress" commits each
roughly representing a day's work over the last six months, that would
take a long time. Mea Culpa. Next time I'll use the WIP commit approach
only for shorter periods and just do the work to merge in and test
changes. The changes in 0.2 are covered in the changelog, which is
updated in this commit.