feat: Support exploding ranged attacks
The ranged attack assessment code already supported explosions, but it
was broken and unused. This commit fixes this code and adds the
necessary additional code to item planning and weapon swapping to
support the Damnation arbalest unrand. Damnation will damage the player
if we shoot at an adjacent monster and can never explode on an empty
square. Hence we need to swap to a non-exploding weapon for adjacent
monsters.

The weapon assessment code has a new "explosion" situation that
de-values exploding weapons, which will allow us to always carry a
backup launcher. The weapon swap plan assesses this situation, swapping
from exploding launchers at a distance of 2 or less. We won't swap back
to any better exploding launcher until the closest monster is at least 5
squares away. This way we don't lose too much damage from swapping
weapons back and forth as monsters move closer.

Finally, give Damnation an extra value of 750, equivalent to the extra
value we give Punk's acid ego. This should allow it to be used over any
mundane weapon.