fix: Missing responses for two prompts
Say Y when attacking allies for the purpose of movement (any checks for
penance will have already been made).

Also say yes to exploration when Zot is near. We are certainly stuck in
this case, but we need to answer Y so that autoexplore can proceed,
finish and trigger a gameplan update.  Otherwise we were having
subsequent movement commands accidentally answer the prompt without ever
actually performing autoexplore. The end result may be the same: we
cycle between levels via the stuck hatch plan. However we want
autoexplore and hence goal updating to work even if we're close to being
punished by the Zot clock.