refactor: Move plan functions to a table
qw has reached 200 local variables, which is the maximum it can have in
the clua environment. This commit begins the process of moving local
variables into tables. The plan cascades are moved into the new `plans`
table.

We also branch-specific plan sets for Abyss into its own cascade, since
these currently need to happen before many other plans, and there's no
appropriate existing cascade for them.

For plans specific to Swamp and Tomb, since the related plan functions
are pretty small and fit naturally into the second exploration cascade,
we just move the functions to plans-explore.lua. The end result is that
the overall move cascade is mostly a set of plan cascades with a few
remaining individual plans.