feat: Better goal movement in the Abyss
When moving towards goal features in the Abyss, allow movement directly
while in danger as long as our HP is at least 75% of max. This way we
don't get so bogged down in combat when we're relatively healthy. Since
I've moved all relevant movement plans to the Abyss plan cascade that
happens before attacking, the logic for moving to abyss objectives is
simplified.

For runelights and sensed abyssal rune, adopt a common strategy of
moving to the reachable position that's closest to the runelight/rune
that's also adjacent to an unexplored square. This approach generally
has us keep moving towards the sensed rune and exploring near it any any
runelights until we find a path to the rune.