feat: Better logic to check for ranged weapons
The previous checks for use of ranged weapons on checked which weapon
skill we used, but it can be the case that we currently don't have a
ranged weapon equip. This would currently only really happen when we get
polymorphed to a bad form that melds weapon. Use a more accurate test
that checks for a weapon and that it's a launcher. The skill-based check
is still used in item planning code.