feat: Better tactics for bad forms
Try harder to flee in bad forms that aren't tree form. Relax conditions
that normally prevent us from fleeing if we're in a bad form, namely
worrying about adjacent monsters or ranged monsters. Also try to
teleport in a bad form if we see we can't flee and don't have
cancellation.