feat: generalize some map movement functions
Have the map movement functions that work for a set of positions call a
function that works for a single position instead of the reverse. This
simplifies the implementations without significant cost to performance.
Make general variants of `best_move_towards_unexplored...` that can find
a reachable unexplored square near an arbitrary position or a closest
reachable unexplored square near a set of positions.

Also rename these distance map functions to drop use of `map_` and
rename any argument to make it clear we're working with map positions
instead of los positions. This makes for shorter function names and
frees up `pos` in these functions as an iterator variable as well.