fix: Re-allow goal updates during the plan cascade
If we don't process a goal update when a turn is not consumed, qw will
proceed through various plans late in the exploration cascade or in the
stuck cascade. This often happens when autoexplore completes and the
level is now considered explored. So specifically for the case when a
plan hasn't consumed a turn and the cascade would continue, do a goal
update if one has been requested.

Complications can arise if we yield during a goal update and then
attempt to send a dummy command when the turn is over, but that
shouldn't be happening for the case being added in this commit.