fix: Improved handling for level transitions
Expire c_persist data for temporary levels immediately after we return
from one, so we don't have things like stale autoexploration data for a
branch or portal that we enter multiple times. The latter has happened
for Bazaars, which can generate multiple times per game.

Improve the logic for detecting when we've arrived at a new level on a
turn, differentiating this from when we've simply loaded a game from a
save.