fix: Properly set branch info for failed travel
When we can't travel into a branch, set `goal_travel.first_branch` in
addition to `goal_travel.stop_branch` so that our list of goal travel
features is populated. This is necessary for travel to understand
that autoexplore should be enabled when we detect a portal but haven't
found it yet.