fix: Correct logic for swamp cloud movement
We broke this a while ago in such a way that qw would always read
through its entire stack of teleport scrolls after getting the Swamp
rune. Ouch! Simplify the logic so that this plan is only concerned with
getting towards the special swamp clouds in order to spot the rune. For
escaping, the distance mapped movement is sufficient.