feat: Simplify distance mapped exploration
Don't have two plans for distance mapped movement to unexplored areas.
Merge the plan that allows movement to an unsafe unexplored square into
the one that moves only to safe unexplored squares. The safe approach is
tried first and we only attempt the unsafe version if this fails. This
all happens after attempting travel, so it will be appropriate to try
these forms of movement in succession if we're otherwise unable to
fulfill exploration goals. It's also the approach used by other similar
distance map movement functions.