feat: Allow use of Fragile equipment
Previously qw refused to use any Fragile item. This commit penalizes
Fragile weapons heavily when they would cut hydra heads and we care
about that (-500 utility, the same as Harm) but will allow their use,
and it doesn't penalize Fragile on other equipment at all. For the
latter, qw only swaps when it finds a superior item, so it's fine if
we'll break a Fragile item in the future in that case.

I've also added a high value for the Consecrated Labrys of 1000. This
will