fix: fleeing and teleport logic for bad forms
Re-order the emergency plans so they're more logical, with "escape"
plans occurring before "buff" plans related to staying and attacking.
The latter will be split into another cascade or moved to the attack
cascade in the future.

Move the teleport plan above the tactical step plans, since we want to
decide this first. Specifically when stuck in a bad form without a
cancellation potion, teleport only if we have adjacent monsters. This
allows us to try to flee if we have sufficient distance.

Also when in bad forms, accept a flee position regardless of monster
distance, since fleeing is still the wisest action. If we don't have
teleport, we'll want the option to flee. Even if we're teleporting, we
can try to move away from nearby monsters until the teleport takes
place, so we take fewer hits.