refactor: Move qw state into a table
This commit is the beginning of moving qw's state into a `qw` table. The
benefits of this are making debugging easier, since all qw state will be
accessible in one table instead of individual local variables, and
making code easier to read, since it will be clear when we're accessing
global state variables.

This first commit moves all of the high level control state into the qw
table, leaving the remaining game logic state to moved in gradually.