fix: travel stair finalization logic
Currently when finalizing travel of a target level, if we're missing
stairs in a given direction, we first search for stairs on the level in
that direction but for stairs in the opposite stair direction. This
created a problem specifically if doing so caused us to enter a
disconnected area on the target level and find the last staircase in our
desired direction when the level in our given direction is missing
stairs in the opposite direction.

In this case, now our target level is explored, the level in our given
direction becomes the new target, and our former target level has
unexplored stairs in our original direction that needs exploration.
Being in a disconnected cell means that we'll fail to go to explored
stairs on the former target level and be forced to execute the stair
backtracking plan. This will lead us to take stairs back to the level in
our original given direction, which is good, but then the former target
level is selected by travel finalization since it's in the new search
direction of our new target. This means we travel immediately back to
our disconnected cell and try to find unexplored stairs, which bad,
because we're stuck in a loop.

The solution of this commit is to search on the level in the given
direction from our target level only if it's not yet autoexplored.
Otherwise we search stairs in the given direction on our target level
before trying stairs in the opposite direction on the level in our given
direction. Now landing in a disconnected area and finding all stairs for
our target level won't conflict with the backtracking plan, since we'll
always prioritize stair exploration on the current target level, which
lets us get to the other disconnected areas on the former target.

The interactions between travel finalization and the stair exploration
resets of the backtracking plan are hard to understand. Ideally we'd use
distance mapping to check stair connectivity and reason about which
specific stairs we need to explore and how we need to get to them.
Having this logic would remove the need for the stair backtracking plan
at all. This is doable, but will take a bit of work, so I'm making this
fix for now.