feat: Improve goal travel finalization
In 015441d9 I tweaked the logic to prevent qw getting stuck in
situations involving disconnected areas where we needed to travel to
another disconnected area on the level in order to finish exploring
stairs. This fix prioritizes exploring stairs in the given exploration
direction on the target level before trying the stairs in the opposite
direction on the adjacent level in the original direction, unless that
adjacent level isn't fully autoexplored. The last aspect will have qw
fully explore e.g. Orc:2 before trying to explore the upstairs on that
level. In general it's better to explore those upstairs and clear
monsters at the destination area sooner so they become safe points where
we can flee to. If we end up taking those upstairs for the first time
when we're trying to flee, fleeing will be much less safe.

Hence this commit tweaks the logic again to do a stair dir search on the
level in the given direction as soon as we see we've found an unexplored
staircase. This only applies if that level is not fully autoexplored, so
that the logic changed in 015441d9 will still work to prevent loops.