fix: Portal expiration problems
Currently if a portal expires in LOS, qw gets stuck, since it only
recognizes the out-of-los expiration messages and never expires the
portal. This commit updates the map code to check for expired portal
features and remove any portal that we have recorded at that location.

I've also tweaked the handling of the generic out-of-los message for an
expired portal. These messages did trigger portal data removal, but
there had to be special check to avoid handling this if a level had
multiple timed portals. With this commit, we add each portal type's
final message to the portal branch data, watch for these messages, and
record each portal that's about to expire in a list. Hence when we
receive a generic message expiration message, we know to expire the
first portal added to the list.