feat: Logic for aiming directly at ranged targets
qw has support for determining when it should fire directly at a target
instead of using the full range bolt. Reasons for doing this include
cases where we would hit friendly monsters past the intended target due
to a miss or when thrown ammunition could end up in deep water or lava
after a miss. However the implementation wasn't working due to bugs and
the fact that spells.path() on the clua side always aimed directly at
the given target. With crawl@1ebeae05, we have an argument for
spells.path() we can use to disable aiming at the target. This commit
uses that argument and fixes the logic on qw's side.