fix: Melee target interface a caching
Like we now do in the ranged attack code, return the attack result table for
melee instead of just the position. Although we currently don't need
this data, ranged and melee share a lot of same assessment functions,
and those consider the returned 'target' to be that result table.

Also fix the caching of the melee target when no target is found.