fix: Corrections to fleeing logic
This commit fixes a bug with exclusions where incoming monsters already
in range to melee us were not seen as not being able to reach us,
leading to them being incorrectly excluded when we got to low HP.

Various checks for player and monster pathing are refactored and renamed
for better clarity as to what they check. The implementation was hard to
follow due to the distance calculations enforcing a condition about
being better than the player's current location. But this isn't
necessary to have the logic work if we just directly compare distances
between the squares we assess.