feat: argument support for the turn memo function
This allows memoizing functions that take simple arguments. Each
argument results in a new nested table the memos variable. In theory it
works with any type of argument value. However it shouldn't be used in
practice with objects like tables where two tables with identical values
won't be seen as the same, since they are different references.