fix: Don't reset distance maps after exploration
Distance map resets should only happen if a cell became
untraversable/excluded and the cell used to be traversable/unexecluded,
not if the cell was previously unseen and the values were nil. The
result of this bug was that distance maps were resetting whenever we
revealed previously unseen terrain. Oops!