feat: Retreating to better attack positions
This implements a retreat system where qw will assess if number of
monsters in LOS that can melee attack is high and if so fall back to the
best nearby position that sufficiently reduces the number of monsters
that can melee us at once.

This system works as part of tactical stepping in cases where we have
monsters adjacent to us, where it will only retreat to a position in LOS
radius. If we don't have adjacent monsters, it will search as far as the
nearest flee position or the entire map if we don't have one.

The full details of the retreat algorithm are described here:

https://github.com/crawl/qw/wiki/Retreating-Tactics

qw has won several GrHu games using this system, so it's reasonably
effective and well-tested at this point. It does need more tuning, since
there are still observed cases where it retreats for marginal gain while
taking ranged damage.