fix: Corrections to monster pathing checks
This fixes a bug with exclusions where incoming monsters already in
range to melee us were not seen as not being able to reach us, leading
to them being incorrectly excluded when we got to low HP.

Various checks for player and monster pathing are refactored and renamed
for better clarity as to what they check. The monster search calculation
code is generalized to allow better returning of search results besides
just the final position, as this will be useful in future work.