feat: Surrounded behaviour when using vamp axes
This commit checks monsters in los that are susceptible to vampirism
when we have this ego on an axe. If there are at least 4 such monsters,
we prefer being surrounded and don't retreat. This may need refinement
as to exactly how many monsters we require and at what radius we check.
Monsters closer to us are more likely to be relevant for vampirism
sustaining a fight, but further away monsters do eventually reach us as
well.