feat: Support for Okawaru's capstone gifts
Now that Okawaru has capstone gifts for weapons and armour as of
crawl@bc85d149, use the lua api support added in crawl@f6de2f5a to use
these abilities. For weapons, we have to decide which of the offered
weapons is best, so we take the weapon with the best current value.

For armour, we have to do something different than what qw has done in
the past for item evaluation, because we're not necessarilly comparing
among items for the same slot. A piece of body armour might offer a
marginal increase in utility compared to the current body armour, while
e.g. boots might offer a larger increase in utility. Simply finding the
armour with the highest utility will overvalue body armour due to the
latter's high base AC etc. I've gone with an approach that looks at the
utility difference between each offered item and the utility of current
item in the corresponding slot, if any. The item with the biggest
difference is selected for armour acquirement.