feat: Add debug statements for acquirement
An "items" debug channel that prints the choices offered by the three
types of acquirement, along with any relevant evaluation of each item.

The dsay() function gets a do_note argument that adds the text to the
note log, which we use for acquirement debug, since these are rare
events where we'd like to check the dump or morgue to see the results.