feat: Move to better terrain for acquirement
Check when our current location would destroy any Okawaru acquirement
gifts and attempt to move to suitable terrain if necessary. It's ideal
to go to the effort of moving instead of simply delaying receiving gifts
since we may travel and end up using piety and losing access to the
gifts. Also employ this for reading acquirement scrolls so we can read
them sooner instead of just holding onto them.