feat: Rework conditions for using various buffs
Currently qw has two lists of monsters to determine when it used buffs:
a list of "scary" monsters to check for using lower-level buffs like
Hero and Berserk a list of "nasty" monsters to check for using
higher-level buffs like Finesse. These lists check specific player XLs,
resists, or a few other aspects (like hydra head chopping) for each
monster entry. This system is flexible, but it's difficult to keep
maintained and accurate, and has generally lead qw to overuse the
lower-level buffs, which tend to cost more piety now than they did in
the past.

This commit replaces the dual monster list system checks of monster
threat level along with a much smaller single list. The new smaller list
contains only the entries checking for monsters versus specific
resistances or other uniquely dangerous aspects, and none of purely
xl-based entries.

We consider a monster in a given radius to be scary whenever it has an
entry in the scary monster list or has a threat level of 3.
Additionally, we score each monster in the radius by threat level,
considering the monsters scary if the total score is 10 or higher (e.g.
5 or more monsters with threat level 2). We use Finesse when we have
  scary monsters, using Hero only when Finesse is not available or when
the monster score is 15 or higher, where Hero is used in addition to
Finesse.

Some very preliminary testing of GrHu on a specific seed that made it to
Zot shows promising results in that qw was able to replay through all of
the seed I've run so far (through Depths:2) and additionally was able to
receive it's Okawaru capstone gifts on Lair:5. This will likely need
further tuning, possibly including different conditions for melee and
we'll need a better condition for a few god abilities like Brothers in
Arms.