feat: Emergency potion and scroll quaff id
We now track identification of potions and scrolls by subtype in
c_persist and use this information to determine if we should quaff id
during combat if our desired potion is unavailable. Currently we only do
this for curing and heal wounds potions, but usage could be exanded to
other potions and scrolls.