feat: Consider traps to be traversable
Having these as untraversable was causing pathing problems on Zot:5.
It's not a good way to deal with avoiding traps, since it will prevent
us from trying to walk over e.g. teleport and dispersal traps in cases
were we must proceed, regardless of safety. The distance mapping
movement code does consider safety and tries to avoid stepping on traps
if possible, so the untraversable approach shouldn't be necessary. If it
turns out that there are still issues with stepping on traps when we
could avoid them, those cases can hopefully be handled in another way.

We also no longer mark web traps as safe, since there should no longer
be issues with this preventing pathing checks from working in web-heavy
areas like Spider.