feat: Limit retreat range versus ranged attackers
When qw retreats from monsters like orbs of fire that have powerful
ranged attacks, it chooses the best retreat location even if it's at a
long-range distance. This means qw often takes heavy fire as it tries to
move to a distant location. Even worse for faster monsters, we'll never
reach that distant location before the monster reaches us. This commit
limits retreat distance when facing such monsters to an in-LOS location
via tactical step. We hence treat non-reaching ranged monsters the same
way we treat monsters that are able to melee us. Monsters that have
reach are assessed for melee already in this regard, and are excluded
from this new ranged check.
currently retreated from to long-range distances. This means qw