feat: Improvements to threat assessment
Instead of the raw threat level, we now use a threat adjusted by factors
such as whether the monster or player are hasted, mighted, berserked,
finessed, or slowed. The total threat score for using heroism when we're
finessed is adjusted accordingly, and we extend the range ot
consideration for that case to all of LOS.

Additionally we consider whether a monster can actually successfully
attack us before adding their threat to a threat score. This is a simple
check for either being a proper ranged monster (i.e. ignoring reach) or
if the monster has a path to reach us for its melee range.