feat: Improve handling of buff potions and slowing
This commit gives better conditions for using haste potions. Namely we
use haste potions if we worshiping Okawaru but don't have either Finesse
or Heroism available, and we haste if we're slowed and the total monster
score is high enough. Haste conditions for other gods when key abilities
are not available will be forthcoming. Before expanding use of haste, I
want to give qw support for using wands so that using haste becomes a
last resort, given its rarity.

I've added freezing wraiths as a scary monster for an appropriate XL
based on a check as to whether we can be slowed. Additionally we now
correctly handle torpor snail aura in terms of whether we need to rest,
and also rest off half-willed status whenever we're not in Tartarus.

Finally, we tidy up various emergency plan functions for potions and god
abilties and more logically order certain plans in the emergency plan
cascade.