refactor: Move global data into the qw table
Move values for various RC variables as well as the version string into
the qw table. For RC variables, use the allcaps global variables at init
time to set up a variable with a lowercase name in qw. These are easier
to type and faster to access (since the qw table is local).

The qw version and configuration notes are now mode whenever qw
initializes, not only at the start of the game. This way we can see when
the qw version and these RC variables change if a game is saved and
resumed.