feat: A general plan to move towards runes
This adds a plan to move towards a rune that can be used in non-Abyss
branches. In case the rune gets excluded or hidden behind firewood, this
plan allows qw to still pick up the rune. The Abyss still has its own
nearly identical rune movement plan, since it needs to be at a higher
priority, in addition to checking its own condition.