feat: Menu selection for various scrolls
For scrolls of identify, branch weapon, enchant weapon, and enchant
armour, we use the newly added clua menu hooks to handle these menus.
This lets us not have to do hacky item slot swaps to prevent wasting
scrolls (in the case of enchant and brand weapon) or waste the scroll
altogether (in the case of identify).

I've also update the list of "good" body armour to include troll leather
armour.