feat: Limit retreat distance based on threat
This commit has qw limit the distance its willing to retreat based on
the total threat in LOS and the difference in the number of enemies able
to melee it at the new position compared to the current one. The aim is
to prevent qw from running past multiple monsters for small improvements
in its position.  The retreat distance limit is:

difference_in_melee_enemies * (threat - ranged_threat / 2)

so that ranged monsters contribute only half as much compared to
non-ranged monsters. Hance qw won't retreat nearly as far in the
presence of many dangerous ranged monsters.