feat: Improved handling of safe monsters
Currently, when a monster in LOS can't even attack us, we don't add it
to the enemy list. This makes it difficult to attack these monsters.
This commit adds them to the enemy list so that they can be attacked,
but records when all monsters in LOS are safe, thus also allowing us to
use rest plans in these situations.