feat: Prioritize attacking los-based threats
For certain monsters like moths of wrath, we'd like to target them as
soon as possible even ignoring whatever is closer to us. This commits
has monsters that return true for :los_danger() be attacked first.
Currenty this list only contains moths of wrath and glass eyes, and doom
hounds that have not yet successfully used their Doom Howl, but it can
be expanded in the future.